+++ variables.tf
@@ -321,4 +321,9 @@ variable "green_header_value" {
   default     = "green"
 }
 
+variable "alb_priority_green" {
+  description = "Priority of ALB rule https://docs.aws.amazon.com/elasticloadbalancing/latest/application/load-balancer-listeners.html#listener-rules"
+  type        = string
+  default     = "110"
+}